Refactor MCP compression code into modules
